MB-5883 Move contact info validation schema to util file
This will enable re-use in other components
